@@ -8,16 +8,17 @@
     "license": "BSD-3-Clause",
     "type": "library",
     "require": {
-        "alex-kalanis/kw_mapper": ">=2.5 <3",
-        "php": ">=7.3.0"
+        "alex-kalanis/kw_mapper": ">=3.0 <4",
+        "php": ">=7.4.0"
     },
     "require-dev": {
-        "alex-kalanis/kw_storage": ">=4 <5",
+        "alex-kalanis/kw_storage": ">=5.0 <6",
         "ext-pdo": "*",
+        "shipmonk/composer-dependency-analyser": "^1.4",
         "friendsofphp/php-cs-fixer": "^3.0",
         "phpstan/phpstan": "^1.0",
         "phpstan/phpstan-phpunit": "^1.0",
-        "phpunit/phpunit": ">=7.2 <=9"
+        "phpunit/phpunit": ">=8.0 <=9"
     },
     "suggest": {
         "ext-pdo": "Content stored in database",
